requirements.txt
twisted
attrs
